requirements.txt
twisted
attrs
